requirements.txt
networkx
matplotlib
pillow
tkinter-tooltip